created gathering system from top master elo

+++ src/main/java/lol/kv3rk/draft_predict/ClientApplication/DTO/MostBannedChampions.java
@@ -3,5 +3,5 @@ package lol.kv3rk.draft_predict.ClientApplication.DTO;
 public interface MostBannedChampions {
     String getChampion();
 
-    int getBanRate();
+    double getBanRate();
 }
